#1363 Refactor transaction.xa.wrap => transaction.xa.convert
