+++ sharding-transaction/sharding-transaction-xa/src/main/java/io/shardingsphere/transaction/xa/convert/XADataSourceMapConverter.java
@@ -15,14 +15,14 @@
  * </p>
  */
 
-package io.shardingsphere.transaction.xa.wrap;
+package io.shardingsphere.transaction.xa.convert;
 
 import io.shardingsphere.core.constant.DatabaseType;
 import io.shardingsphere.spi.transaction.xa.DataSourceMapConverter;
 import io.shardingsphere.transaction.manager.xa.XATransactionManager;
 import io.shardingsphere.transaction.xa.manager.XATransactionManagerSPILoader;
-import io.shardingsphere.transaction.xa.wrap.dialect.XADataSourceFactory;
-import io.shardingsphere.transaction.xa.wrap.extractor.DataSourceParameterFactory;
+import io.shardingsphere.transaction.xa.convert.dialect.XADataSourceFactory;
+import io.shardingsphere.transaction.xa.convert.extractor.DataSourceParameterFactory;
 
 import javax.sql.DataSource;
 import java.util.HashMap;
