+++ sharding-transaction/sharding-transaction-xa/src/main/java/io/shardingsphere/transaction/xa/manager/XATransactionDataSourceWrapper.java
@@ -21,8 +21,8 @@ import com.atomikos.beans.PropertyException;
 import com.atomikos.beans.PropertyUtils;
 import com.atomikos.jdbc.AtomikosDataSourceBean;
 import io.shardingsphere.core.rule.DataSourceParameter;
-import io.shardingsphere.transaction.xa.wrap.dialect.XADatabaseType;
-import io.shardingsphere.transaction.xa.wrap.dialect.XAPropertyFactory;
+import io.shardingsphere.transaction.xa.convert.dialect.XADatabaseType;
+import io.shardingsphere.transaction.xa.convert.dialect.XAPropertyFactory;
 import lombok.RequiredArgsConstructor;
 import org.apache.tomcat.dbcp.dbcp2.managed.BasicManagedDataSource;
