+++ sharding-transaction/sharding-transaction-xa/src/main/resources/META-INF/services/io.shardingsphere.spi.transaction.xa.DataSourceMapConverter
@@ -1,2 +1,2 @@
-io.shardingsphere.transaction.xa.wrap.XADataSourceMapConverter
+io.shardingsphere.transaction.xa.convert.XADataSourceMapConverter
