+++ sharding-transaction/sharding-transaction-xa/src/test/java/io/shardingsphere/transaction/xa/convert/dialect/XAPropertyFactoryTest.java
@@ -15,7 +15,7 @@
  * </p>
  */
 
-package io.shardingsphere.transaction.xa.wrap.dialect;
+package io.shardingsphere.transaction.xa.convert.dialect;
 
 import io.shardingsphere.core.rule.DataSourceParameter;
 import org.junit.Before;
