+++ sharding-transaction/sharding-transaction-xa/src/test/java/io/shardingsphere/transaction/xa/manager/XATransactionDataSourceWrapperTest.java
@@ -22,9 +22,9 @@ import com.atomikos.jdbc.AtomikosDataSourceBean;
 import io.shardingsphere.core.constant.DatabaseType;
 import io.shardingsphere.core.constant.transaction.ProxyPoolType;
 import io.shardingsphere.core.rule.DataSourceParameter;
-import io.shardingsphere.transaction.xa.wrap.dialect.XADataSourceFactory;
+import io.shardingsphere.transaction.xa.convert.dialect.XADataSourceFactory;
 import io.shardingsphere.transaction.xa.fixture.ReflectiveUtil;
-import io.shardingsphere.transaction.xa.wrap.dialect.XADatabaseType;
+import io.shardingsphere.transaction.xa.convert.dialect.XADatabaseType;
 import org.apache.tomcat.dbcp.dbcp2.managed.BasicManagedDataSource;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
